tdf#125600 hunspell patch: don't suggest itself as spelling correction
The example was kHz. The user wrongly spelled it as KHz.
But suggestions are always capitalized for a capitalized misspelling,
so correct suggestion kHz was being capitalized back to KHz.

Don't suggest a capitalized form of a correct suggestion
that exactly matches the wrongly-capitalized word being replaced.

if capwords: then the misspelled word is in caps (perhaps all caps)
mkinitsmall: now the misspelled word is first-char-lower-cased
if initsmallword == j: correct lowercase would match incorrect upper

Change-Id: I0622158ebf2cb65f070547e061f5136df0564051
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191755
Reviewed-by: László Németh <nemeth@numbertext.org>
Tested-by: Jenkins